feat: footer 모바일/데스크톱 분리 및 모바일 UI 추가
